clean up tools + py licenses

+++ src/shared/qtgui/qtlua.cpp
@@ -2741,7 +2741,7 @@ int qslider_new(lua_State *L)
     if (argc < 0 and argc > 1) {
        error(fn, 0);
     } else {
-        int arg0 = (argc >= 1) ? lua_tointeger(L, 1) : (int) Qt::Vertical;
+        int arg0 = (argc >= 1) ? lua_tointeger(L, 1) : Qt::Vertical;
         return addWidget(new QSlider(static_cast<Qt::Orientation>(arg0)));
     }
     return 0;
